@@ -4,68 +4,45 @@ declare(strict_types=1);
 
 namespace App\Controller;
 
-use Doctrine\DBAL\Connection;
-use Doctrine\DBAL\FetchMode;
+use App\Repository\FlightInformationRepository;
 use Knp\Component\Pager\Event\Subscriber\Paginate\Callback\CallbackPagination;
+use Knp\Component\Pager\Pagination\PaginationInterface;
 use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-final class IndexController extends AbstractController
+class IndexController extends AbstractController
 {
-    /**
-     * @var \Doctrine\DBAL\Connection
-     */
-    private $connection;
+    private FlightInformationRepository $repository;
+    private PaginatorInterface $paginator;
 
-    /**
-     * IndexController constructor.
-     * @param \Doctrine\DBAL\Connection $connection
-     */
-    public function __construct(Connection $connection)
+    public function __construct(FlightInformationRepository $repository, PaginatorInterface $paginator)
     {
-        $this->connection = $connection;
+        $this->repository = $repository;
+        $this->paginator = $paginator;
     }
 
     /**
-     * @Route("/", name="index", methods="GET")
-     * @param \Knp\Component\Pager\PaginatorInterface $paginator
-     * @param \Symfony\Component\HttpFoundation\Request $request
-     * @return \Symfony\Component\HttpFoundation\Response
+     * @Route("/", name="main", methods="GET")
      */
-    public function index(PaginatorInterface $paginator, Request $request): Response
+    public function main(Request $request): Response
     {
-        $count = function () {
-            $queryBuilder = $this->connection->createQueryBuilder();
-            $query = $queryBuilder
-                ->select('*')
-                ->from('departures');
-            $statement = $query->execute();
-            return $statement->rowCount();
-        };
-
-        $items = function ($offset, $limit) {
-            $queryBuilder = $this->connection->createQueryBuilder();
-            $query = $queryBuilder
-                ->select('*')
-                ->from('departures')
-                ->setFirstResult($offset)
-                ->setMaxResults($limit);
-            $statement = $query->execute();
-            return $statement->fetchAll(FetchMode::ASSOCIATIVE);
-        };
-
-        $target = new CallbackPagination($count, $items);
-
         $page = $request->query->getInt('page', 1);
 
-        $pagination = $paginator->paginate($target, $page, 6);
-
         return $this->render('index/index.html.twig', [
-            'pagination' => $pagination,
+            'pagination' => $this->createPagination($page, 6),
         ]);
+    }
 
+    private function createPagination(int $page, int $limit): PaginationInterface
+    {
+        $target = new CallbackPagination(
+            fn() => $this->repository->countFlightInformation(),
+            fn($offset, $limit) => $this->repository->findFlightInformationInterval($offset, $limit)
+        );
+
+        return $this->paginator->paginate($target, $page, $limit);
     }
 }
